requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
alsa-lib=1.2.3=h516909a_0
aom=3.3.0=h27087fc_1
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py38h497a2fe_1
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.10.0=pyha770c72_0
black=22.8.0=pypi_0
blas=2.114=mkl
blas-devel=3.9.0=14_linux64_mkl
bleach=4.1.0=pyhd8ed1ab_0
blosc=1.21.0=h9c3ff4c_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py38h0a891b7_1004
brunsli=0.1=h9c3ff4c_0
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
c-blosc2=2.0.4=h5f21a17_1
ca-certificates=2023.7.22=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.0=py38h3931269_0
cfitsio=4.1.0=hd9d235c_0
charls=2.3.4=h9c3ff4c_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.3=pypi_0
cloudpickle=2.0.0=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
configparser=5.2.0=pyhd8ed1ab_0
cryptography=36.0.2=py38h2b5fc30_1
cudatoolkit=11.3.1=ha36c431_10
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.11.2=py38h0a891b7_2
dask-core=2022.4.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=h5008d03_3
debugpy=1.5.1=py38h709712a_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
easydict=1.9=py_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
expat=2.4.8=h27087fc_0
ffmpeg=4.3=hf484d3e_0
fire=0.4.0=pypi_0
flake8=4.0.1=pyhd8ed1ab_2
flit-core=3.7.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h8e229c2_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.31.2=py38h0a891b7_1
freetype=2.10.4=h0708190_1
fsspec=2022.3.0=pyhd8ed1ab_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gst-plugins-base=1.20.1=hcf0ee16_1
gstreamer=1.20.1=hd4edc92_1
h5py=3.6.0=nompi_py38hfbb2109_100
hd-bet=1.0=dev_0
hdf5=1.12.1=nompi_h2386368_104
icu=69.1=h9c3ff4c_0
idna=3.3=pyhd8ed1ab_0
imagecodecs=2022.2.22=py38he865337_1
imageio=2.16.1=pyhcf75d05_0
importlib-metadata=4.11.3=py38h578d9bd_1
importlib_resources=5.6.0=pyhd8ed1ab_0
ipdb=0.13.9=pyhd8ed1ab_0
ipykernel=6.12.1=py38h7f3c49e_0
ipympl=0.8.8=pyhd8ed1ab_0
ipython=8.2.0=py38h578d9bd_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jbig=2.1=h7f98852_2003
jedi=0.18.1=py38h578d9bd_1
jinja2=3.1.1=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h7f98852_0
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter_client=7.2.1=pyhd8ed1ab_0
jupyter_core=4.9.2=py38h578d9bd_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
jxrlib=1.1=h7f98852_2
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py38h43d8883_1
krb5=1.19.3=h3790be6_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libaec=1.0.6=h9c3ff4c_0
libavif=0.9.3=h166bdaf_1
libblas=3.9.0=14_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=14_linux64_mkl
libclang=13.0.1=default_hc23dcda_0
libcurl=7.82.0=h7bff187_0
libdeflate=1.10=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=11.2.0=h69a702a_14
libgfortran5=11.2.0=h5c6108e_14
libglib=2.70.2=h174f98d_4
libiconv=1.16=h516909a_0
liblapack=3.9.0=14_linux64_mkl
liblapacke=3.9.0=14_linux64_mkl
libllvm13=13.0.1=hf817b99_2
libnghttp2=1.47.0=h727a467_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=14.2=hd57d9b9_0
libprotobuf=3.20.0=h6239696_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_14
libtiff=4.3.0=h542a066_3
libuuid=2.32.1=h7f98852_1000
libuv=1.43.0=h7f98852_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h885dcf4_1
libzlib=1.2.11=h166bdaf_1014
libzopfli=1.0.3=h9c3ff4c_0
llvm-openmp=13.0.1=he0ac6c6_1
locket=0.2.0=py_2
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.1.1=py38h0a891b7_1
matplotlib=3.5.1=py38h578d9bd_0
matplotlib-base=3.5.1=py38hf4fb855_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mccabe=0.6.1=py_1
medmnist=2.1.0=pypi_0
mistune=0.8.4=py38h497a2fe_1005
mkl=2022.0.1=h8d4b97c_803
mkl-devel=2022.0.1=ha770c72_804
mkl-include=2022.0.1=h8d4b97c_803
munkres=1.1.4=pyh9f0ad1d_0
mypy-extensions=0.4.3=pypi_0
mysql-common=8.0.28=haf5c9bc_3
mysql-libs=8.0.28=h28c427c_3
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.5=pyhd8ed1ab_2
nbconvert-core=6.4.5=pyhd8ed1ab_2
nbconvert-pandoc=6.4.5=pyhd8ed1ab_2
nbformat=5.3.0=pyhd8ed1ab_0
ncurses=6.3=h9c3ff4c_0
nest-asyncio=1.5.5=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
networkx=2.7.1=pyhd8ed1ab_1
nibabel=3.2.2=pyhd8ed1ab_0
notebook=6.4.10=pyha770c72_0
nspr=4.32=h9c3ff4c_1
nss=3.77=h2350873_0
numpy=1.22.3=py38h05e7239_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1n=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py38h47df419_0
pandoc=2.17.1.1=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
pathspec=0.10.1=pypi_0
pathtools=0.1.2=py_1
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.1.0=py38h0ee0e06_0
pip=22.0.4=pyhd8ed1ab_0
platformdirs=2.5.2=pypi_0
prometheus_client=0.14.0=pyhd8ed1ab_0
promise=2.3=py38h578d9bd_5
prompt-toolkit=3.0.29=pyha770c72_0
protobuf=3.20.0=py38hfa26641_3
psutil=5.9.0=py38h0a891b7_1
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycodestyle=2.8.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pydicom=2.3.0=pyh6c4a22f_0
pyflakes=2.4.0=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyqt=5.12.3=py38h578d9bd_8
pyqt-impl=5.12.3=py38h0ffb2e6_8
pyqt5-sip=4.19.18=py38h709712a_8
pyqtchart=5.12=py38h7400c14_8
pyqtwebengine=5.12.1=py38h7400c14_8
pyrsistent=0.18.1=py38h0a891b7_1
pysocks=1.7.1=py38h578d9bd_5
python=3.8.13=h582c2e5_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.2=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pywavelets=1.3.0=py38h3ec907f_0
pyyaml=6.0=py38h0a891b7_4
pyzmq=22.3.0=py38hfc09fa9_2
qt=5.12.9=h1304e3e_6
readline=8.1=h46c0cb4_0
requests=2.27.1=pyhd8ed1ab_0
scikit-image=0.19.2=py38h43a58ef_0
scikit-learn=1.0.2=py38h1561384_0
scipy=1.8.0=py38h56a6a73_1
send2trash=1.8.0=pyhd8ed1ab_0
sentry-sdk=1.5.8=pyhd8ed1ab_0
setuptools=62.0.0=py38h578d9bd_0
shortuuid=1.0.8=py38h578d9bd_0
simpleitk=2.2.1=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=3.0.5=pyh44b312d_0
snappy=1.1.8=he1b5a44_3
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.37.1=h4ff8645_0
stack_data=0.2.0=pyhd8ed1ab_0
subprocess32=3.5.4=py_1
tbb=2021.5.0=h924138e_1
termcolor=1.1.0=py_2
terminado=0.13.3=py38h578d9bd_1
testpath=0.6.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2022.3.25=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
tomli=2.0.1=pypi_0
toolz=0.11.2=pyhd8ed1ab_0
torchvision=0.11.3=py38_cu113
tornado=6.1=py38h0a891b7_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
unicodedata2=14.0.0=py38h0a891b7_1
unzip=6.0=h7f98852_3
urllib3=1.26.9=pyhd8ed1ab_0
wandb=0.12.10=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.0=py38h578d9bd_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
yaml=0.2.5=h7f98852_2
yaspin=2.1.0=pyhd8ed1ab_0
zeromq=4.3.4=h9c3ff4c_1
zfp=0.5.5=h9c3ff4c_8
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h166bdaf_1014
zstd=1.5.2=ha95c52a_0
